requirements.txt
# Core dependencies
boto3>=1.34.0
botocore>=1.34.0
aws-cdk-lib>=2.100.0
constructs>=10.3.0

# Data processing and ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
prophet>=1.1.4
statsmodels>=0.14.0

# AWS SDKs and tools
aws-lambda-powertools>=2.30.0
aws-xray-sdk>=2.12.0
mangum>=0.17.0

# Web framework and API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
httpx>=0.25.0

# Database and time series
pytz>=2023.3
python-dateutil>=2.8.2

# Configuration and utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.7.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
moto>=4.2.0

# Development
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Async support
asyncio-mqtt>=0.16.0
aiohttp>=3.9.0

# Logging and monitoring
structlog>=23.2.0
prometheus-client>=0.19.0
